Harden demo/live OKX coverage and REST error handling
- promote instIdCode to long across instrument and order request models for demo responses that exceed 32-bit values
- add EVENTS instrument type support and expand public contract tests to production and demo snapshots
- split committed live fixtures into Production and Demo datasets and capture all public instruments variants plus borrow history
- fix top-level OKX code/msg handling for generic REST list/model responses so API errors are no longer reported as successful calls
- update integration tests, fixture capture docs, and release notes for the new validation flow

+++ OKX.Api/Public/Responses/OkxPublicInstrument.cs
@@ -21,9 +21,10 @@ public record OkxPublicInstrument
     /// Instrument ID code.
     /// For simple binary encoding, you must use instIdCode instead of instId.
     /// For the same instId, it's value may be different between production and demo trading.
+    /// Demo trading responses may return values that exceed 32-bit integer range.
     /// </summary>
     [JsonProperty("instIdCode")]
-    public int? InstrumentIdCode { get; set; }
+    public long? InstrumentIdCode { get; set; }
 
     /// <summary>
     /// Underlying, e.g. BTC-USD. Only applicable to FUTURES/SWAP/OPTION
